command line parameters, filter and sort reports by timestamp, select keyfiles by prefix
Since apple's servers seem to send all the reports regardless of startDate
and endDate, I'm only showing the results after the timestamp specified on
the commandline.
Using the prefix option (to get all reports from the same tag
that's using more than one key) it will add a "key" fields to each
report with the part of the filename after the prefix.
Another minor modification is the key "isodatetime" (easier to
read for an human) and "goog" with an url to see the
location on google maps.